hestiaHUGO: made HTML content rendering as partial function
Since we routed everything to format.html, the content page itself
has to be functionalized in order to render different set of contents
based on the context of the page (e.g. 404, redirect, and etc). Hence,
let's functionalize the content rendering part.

This patch makes HTML content rendering into a partial function.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>