hestiaHUGO: refactored multilingual and multiple formats data processors
Since the key to organize a good multilingual and multiple formats
data in an uniformed and easy to use data structure, we need to
refactor the current ones before develop the rendering partial
functions. Hence, let's do this.

This patch refactors multilingual and multiple formats data processors
in hestiaHUGO/ directory.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>