hestiaHUGO: refactored Hestia partial shortcode url to match latest
The Hestia URL partial shortcode is terribly outdated to the point
where it did not pick up input format and the latest .Formats data
structure keys. Hence, we need to update it before proceeding.
Let's do this.

This patch refactors Hestia partial shortcode url to match the latest
updates in hestiaHUGO/ directory.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>