hestiaHUGO: refactored Length validator to return single console
Since there is a better and cleaner way to return both $dataList
and $console output, we should apply it to all new validator
functions. Hence, let's apply it into the Length dimension validator.

This patch refactors Length validator to return a standardized
manner of dataList and console output in hestiaHUGO/ directory.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>