hestiaHUGO: fixed variables parsing bug for string/__splitArgs p.functions
The partial function string/__splitArgs currently cannot parse
variable context filling due to the absent of the remainder buffer
parsing (post loop). Hence, we need to fix this issue since it is
the critical work to deal with.

This patch fixes variables parsing bug for string/__splitArgs
partial functions in hestiaHUGO/ directory.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>