hestiaHUGO: refactored Get partial function to be premitive
Since we need a safe query of an unknown data structure for Get
against any given data structure, we have to refactor the existing
one that is locked to using Hestia data structure. Hence, let's
refactor it.

This patch refactors Get partial function to be primitive function
and can be used to search any given data structure in hestiaHUGO/
directory.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>