hestiaHUGO: updated PWA .Name.ID to match its processor
Appearently, the PWA ID was meant to append towards the URL of
the start_url or base URL. If we provide an ID that is based on
.URL.Base, we will get a repeating .URL.Base ID. That does not make
any sense. Hence, we sohuld update it to match the proper URL instead.
Let's use current language code since it matches the URL nicely.
Let's do this.

This patch updates PWA .Name.ID to match its processor.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>