hestiaHUGO: added PWA service worker boilerplate
Since PWA service worker javascript script is strongly recommended
to be placed at the root of the web filesystem (as it has scope
limitations for where it sits stated in:
https://web.dev/learn/pwa/service-workers/), we have no choice but
to use Hugo pipe to generate the necessary boilerplate for such
file. Hence, let's do this.

This patch adds PWA service worker boilerplate into hestiaHUGO/
directory.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>