heistaHUGO: upgraded icon filters to be specific
Since to support browserconfig.xml legacy IE download, we have to
generate one of its specific 2:1 sized icon image which is only
used by it. Hence, we have no choice but to upgrade the icon filters
for existing ones to be specific to their requests. Let's do this.

This patch updgrades icon filters to be specific in hestiaHUGO/
directory.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>