hestiaHUGO: refactored banner logo to actual 1200x630 size
Since there is an eed to generate 1200x580.65 due to Microsoft tile
thumbnails, we have to refactor the banner logo up to the largest
format which is 1200x630 where open graph thumbnails used. Then, it
is easier to downsize to the smaller dimensions needed for thumbnail
generations. Hence, let's refactor it.

This patch refactors banner logo to actual 1200x630 size in
hestiaHUGO/ directory.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>