hestiaHUGO: removed redundant language ID field input from user
Since we requested user to construct and input language code AND
have it set as ID explictly, it is better we just ask the user to
input the language code directly as ID and process both code and ID
implictly. Hence, let's do this.

This patch removes redundant language ID field input from user in
hestiaHUGO/ directory.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>