hestiaHUGO: added inline or include first CSS rendering switching feature
Since it is up to the user to decide whether to render inline list
first or include list first, we have to add a setting switch for
them to decide and set a default decision for those who does not
need. Hence, let's do this.

This patch adds inline/include first CSS rendering switching
feature into hestiaHUGO/ directory.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>