hestiaGO: permit the use of MixCaps in package naming
Appearently, MixCaps was opinionated not to be used in package
naming. However, we need it as we're combining hestia- prefix to
package name. Therefore, we will relax this rule for maintenance
and communication sake. Afterall, the goal is to develop intuitive
import clauses and not accepting opinion blindly without actual
reason. Hence, let's relax it.

This patch permit the use of MixCaps in package naming for
hestiaGO in hestiaGO/ directory.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>