hestiaGO: refactored to be absolute independent package
Now, the last piece of dependency is hestiaERROR, let's remove it
as well and refactor the entire hestiaTESTING package to be completely
independent. Hence, let's do this.

This patch refactors hestiaTESTING package to remove all hestiaGO
packages inclusive hestiaERROR package and made it into completely
independent package in hestiaGO/ directory.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>