hestiaGO: shifted hestiaTESTING string verdict data to Data.go
After developing interpret function in hestia_testing module for
hestiaRUST, it is found that the string verdict data was misplaced.
Hence, we should shift it to Data.go aligning to hestia_testing in
hestiaRUST module. Let's do this.

This patch shifts hestiaTESTING string verdict data to Data.go in
hestiaGO/ directory.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>