hestiaGO: backported string slice switches to hestiaTESTING package
Since in hestiaRUST the use of string slice switches had proven
too good for a simple implementation, we should backport it. Moreover,
after the backport, it was found that tinygo is able to operate
properly with the new switches. Hence, let's do this.

This patch backports string slice switches to hestiaTESTING package
in hestiaGO/ directory.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>