hestiaHUGO: enhanced sitemap.xml with self-discovery function
As sitemap.xml has a limit of 50MB or 50k URLs, we have to make
sure it uses an algorithmic approach to reduce the tendency from
hitting that limit close 0 probabilty. Hence, let's enhance the
current version with self-discovery function.

This patch enhances sitemap.xml rendering with self-discovery function
in hestiaHUGO/ directory.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>