hestiaHUGO: refactored console function to be friendly with non-Hestia data
Since there is a tendency of feeding Hestia console with non-Hestia
data structure, we shouldn't yell at the developer who is already
struggling with using Hestia theme already. Hence, let's make it
more friendly to them so that they can help themselves out with a
friendly learning space.

This patch refactors console function to be friendly with non-Hestia
data structure in hestiaHUGO/ directory.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>