hestiaGO: updated hestiaSTRING to match the latest hestiaBITS API
Since we need to develop hestiaSTRING in concurrent with
hestiaINTERNAL/hestiaFMT where hestiaSTRING governs the test codes
of the hestiFMT's formatting functions, we have to update its current
hestiaBITS API to the latest for later development. Hence, let's do
this.

This patch updates hestiaSTRING to match the latest hestiaBITS API
in hestiaGO/ directory.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>