hestiaHUGO: simplify component data processor for a friendlier UI dev
The current component data processor enforces the filename to certain
rules instead of facilitating freedom to UI developer. This is
undesirable and we should simplify it with "as it is" policy. Hence,
let's proceed to refactor it.

This patch simplifies component data processor for a friendlier UI
development in hestiaHUGO/ directory.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>