hestiaHUGO: renamed components to have UI lead name for easier ID
After browsing through the components directory, it was found that
by using the artist name lead, it will be very hard to search for
the component where there is the usual intention. Hence, let's invert
the naming convention to have component-led names and then the artist
name. Hence, let's do this.

This patch renames components to have UI lead name for easier ID
in hestiaHUGO/ directory.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>