hestiaHUGO: refactored hestia shortcode meta-processing function
The existing hestia shortcode is incomplete due to its infancy. Now
that we need to develop a data query shortcode, we have to grow the
meta processing function to a mature level. Hence, let's do this.

This patch refactors hestia shortcode meta-processing function to
a mature level for use in hestiaHUGO/ directory.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>