root: re-organized docs/ directory as a releasable product
Since we want to do things profressionally and ZORALab's HestiaHUGO
is capable of self-relying on documentation part, we should re-organize
ZORALab's Hestia documentation directory as a releasable product so
that we don't have to maintain various versions of documentations just
to be backward-compatiable. Hence, let's do this.

This patch re-organizes docs/ directory as a releasable product in
root/ directory.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>