hestiaHUGO: added Hestia checker for Page Titles hugo data processing fx.
Since the checker is falls outside out Hestia's checking function
when performing a string initialization, it's better we add a checker
function for it for capturing and presenting the error in Hestia way.
Hence, let's do this.

This patch adds Hestia checker for Page Titles hugo data processing
function in hestiaHUGO/ directory.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>