hestiaHUGO: restored capabiltiy of using .Titles.Current in pre-processing
The current hugo .Titles data processing for .Page is very restrictive
to raw data query. To avoid having folks to mess with the language
selection every single time, we have to enable the .Titles.Current
processing before .Titles.Page and for it to processing its data.
The feature went missing during one of the development optimization
so we have to restore it. Hence, let's do this.

This patch restores capability of using .Titles.Current for
.Titles.Page pre-processing in hestiaHUGO/ directory.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>