hestiaHUGO: fixed page-level data parsing bug
The page-level data parsing capability was impaired due to an
opitmization done in the development. Basically the cause is that
the hestiaHUGO/archetypes was outdated and the parsing search
function fails to pick up the newly updated .Data.Filename field.
Hence, let's go ahead and fix them.

This patch fixes page-level data parsing bug in hestiaHUGO/
directory.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>