hestiaHUGO: added number/Int string linter into Hestia data processor
Appearently, there are requests for checking a string is convertable
to integer before actually converting it. Hence, we have to develop
such linter for it. Let's do this.

This patch adds number/Int string linter into Hestia data processor
function for checking a given string is indeed int-convertable in
hestiaHUGO/ directory.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>