hestiaHUGO: encourages the use of TOML data format for data files
We had been through hell just to create and maintain the .data.json
file while documenting blockquote_hestiaUI specification page. It's
very annoying, not flexible, and error prone. Fortunately, we managed
to pivot it to use TOML data format instead and everything was done
smoothly. Hence, let's encourage the use of TOML over JSON by setting
it as a default.

This patch encourages the use of TOML data format for data files in
hestiaHUGO/ directory.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>