hestiaHUGO: added 2nd string processing function
Since this string processing function is too heavy for iterative
function, we have to continuously develop a new one to gain back
some performances. Hence, let's go ahead and build it.

This patch adds a 2nd string processing function as an iterative
development to improve its performance in hestiaHUGO/ directory.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>