docs: refactored based on test results feedback
Based on test results, it seems that the navigation panel is not
responsive and stayed small while the link are too small to read.
Secondly the i18n redirection is so bad that we have to properly
route each compoennts properly. Hence, let's refactor the documentations.

This patch refactors docs/ based on test results feedback.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>