hestiaHUGO: fixed PWA fetch policy and algorithm
It seems that the PWA service worker's policy and algorithms are
causing some presentation problem. The cache is supposedly be
loading from cache repository and the first expected behavior is
network first - not cache first (offline assets requires concious
efforts). Currently, both network first and cache first policies
are mixed up. Hence, we need to fix them in order to get the system
back in shape. Let's do this.

This patch fixes PWA fetch policy and algorithm problem by setting
each algorithms as it is and default to network-first policy in
hestiaHUGO/ directory.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>