hestiaHUGO: updated __data.toml with commented table
The existing __data.toml in archetype directory is using an active
example table. This is not good for people as we just want to show
it as example and not contribute it as part of dataset. Therefore,
we should comment them out in case of future creation. Hence, let's
update it.

This patch updates __data.toml with commented table in hestiaHUGO/
directory.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>