hestiaHUGO: shifted icons rendering to render stage
Since we don't need to render the icons and favicons repeatedly
for every single page, it's best we shift it to render stage only
when the page is '.IsRoot' flagged. This will save up the unecessary
computing power and thus boost the performance as it should. Hence,
let's do this.

This patch shifts icons rendering to render stage in hestiaHUGO/
directory.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>