hestiaHUGO: set core_hestiaUI to have text left-aligned by default
Based on Sibert's and Pei Sun's feedbacks, left-aligned makes the
entire site a lot easier to read and looks more profressional. While
this is great, it can break features. However, if this is direction
to go, then we should proceed. Hence, let's do this.

This patch sets core_hestiaUI to have text left-aligned by default
in hestiaHUGO/ directory.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>