hestiaHUGO: added data parsing capability for data-type output
The existing data-type output like __content.json is returning a
large string instead of its data structure. This is due to the lack
of data parsing function where Hugo needs it. Hence, let's enhance
the content rendering capability for it.

This patch adds data parsing capability for data-type output in
hestiaHUGO/ directory.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>