hestiaHUGO: permitted go functions usage in __content.ldjson
Since we want to standardize everything, we can now proceed to
refactor LDJSON to be consistent with other using __content.ldjson.
With this patch, LDJSON is now capable of using Go functions
with no problem. Hence, let's do this.

This patch permits go functions usage in __content.ldjson for
hestiaHUGO/ directory.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>