hestiaHUGO: changed all __content.[FORMAT] with .hestia[FORMAT] extension
Hugo has a very stubborn way of processing .html file to the point
we have to make a workaround implicitly. To do so, we prefix Hestia
brand name for each extension type, making it uniquely identified
as ZORALab's Hestia ones. Hence, let's do it.

This patch changes all __content.[FORMAT] with .hestia[FORMAT] extension
for templating and consistency purposes.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>