hestiaHUGO: implemented the correct persistent memory API
The .Scratch was not properly implemented as it is scoped to .Page
context. This is not feasible. What we want is .Store which is the
global listing. With .Store, it will let us eliminate unnecessary
recursive scanning on each re-build. The scan only work once on boot.
Hence, let's implement this.

This patch implements the correct persistent memory API store in
hestiaHUGO/ directory.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>