hestiaHUGO: replaced .Creators with .Contributors
The existing .Creators data structure is not i18n capable, not LDJSON
friendly, and worse, are being processed in every single page and
in every of their metadata elements. This is not friendly and will
cause performance damages. Hence, we need to resolve the problems
by replacing it with .Contributors.

This patch replaces .Creators with .Contributors in hestiaHUGO/
directory.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>