hestiaHUGO: added documentations link for future optimization
All the LD+JSON data renderer were not documented to its offical
site. Hence, we have to do it in case new updates rolled down from
upstream. That way, we can track which and what for the work to do.
Let's do this.

This patch adds documentations link for future optimization.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>