hestiaHUGO: amended PWA's local Protocol complaints
It seems that hard-coding PWA local protocol can cause user to mess
with the PWA default settings every single time. This is not feasible
and in fact, this is an error for Chrome failing to recognize UTF-8
wording. However, to comply with status quo, we have no choice but to
reserve Short name for ASCII wording. Hence, let's do this.

This patch amends PWA's local Protocol complaints by strictly setting
Site's Short name to ASCII only wording.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>