hestiaHUGO: optimized hugo's data processor's critical data path
While porting 3 pages (2 are empty, 1 is under development), it is
duly noted that the processing time went back to ~30 seconds. This
is not feasible so our only option is to abstract the hugo's data
processing critical data path out and reduce its overhead. Hence,
let's do this.

This patch optimizes hugo's data processor's critical data path in
hestiaHUGO/ directory.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>