hestiaHUGO: removed error checking for caching
It does not matter with the console message from the offline
caching function. There is no actionable step to be taken. It also
prevent the server from booting up by raising such error which is
not feasible. Hence, let's remove it.

This patch removes error checking for caching in hestiaHUGO/
directory.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>