hestiaHUGO: shifted Metadata structure clean up back to its processor
The existing Metadata flag checking in the critical path is making
the taks to cumbersome and not functioning at all. This is because
at that time, .IsMetadata is not available at all. Hence, we have
to shift it out and place it back to its general data processor
where it received the critical data and then purge the unnecessary
ones off it. Hence, let's do this.

This patch shifts Metadata structure clean up functions back to its
parent processor.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>