hestiaHUGO: refactored with 2nd version standard library layout
After 1 week of proper scanning, planning, and dropping Rust programming
language from the project, we are now capable to unify the library
layouts with Go, TinyGo, and Hugo. Previously, Rust attempted to
force all other languages to only comply for its naming conventions
and layout or otherwise its compiler shall refuse to function or
set our downstream users to suffer what we been through. This is not
straight up authoritarian goverence, blocking productivitiy, and
completely not ideal.

Moreover, the Rust Foundation has the guts and resources to mess with
trademarks and had shown enough laziness and negligence for not reading
the draft internally (case: prohibits users from calling Rust wordmark).
There are a lot of technical issues are yet to resolve like merging
gcov into primary tool similar to cargo; or merge gcov, cargo, and
rustc into 'rust' similar to go, zig, nim, gcc, etc.

Because of this, we had decided to drop Rust programming langauge
entirely although its offering of memory safety is sweet and nice.
It's too risky and even attempting to anchor the entire project
down. Rust so far only suitables for "rewrite everything in Rust"
approach so we will place it in backbench and KIV it.

Henceforth, we refactor hestiaHUGO to our original plan with our
naming conventions not conflicting with actual standard libraries.
Hence, let's do this.

This patch refactors hestiaHUGO/ with 2nd version standard library
layout.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>