hestiaHUGO: added defensive query mechanism to all data processor APIs
Appearently, the recent refactoring caused the removal of the defensive
query algorithm for proper error reporting. Hence, we need to add it
back especially for those APIs that we're offering to users to process
their data. Let's do this.

This patch adds defensive query mechanism to all data processor APIs
in hestiaHUGO/ directory.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>