hestiaHUGO: fixed hestiaMEDIA HTML rendering bug
There are some bugs releated to hestiaMEDIA's HTML rendering and
data sanitization. This is mainly because the Source.Descriptor
was wrongly saved as Source.Media upon sanitization. When a wrong
value is provided to the wrong field, the generated HTML code will
not work. Hence, we need to fix it.

This patch fixes hestiaMEDIA HTML rendering bug in hestiaHUGO/
directory.

Co-authored-by: Galyna, Cory <cory.galyna@zoralab.com>
Co-authored-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>
Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>