hestiaHUGO: fixed hugo compilation bug for PWA manifest and worker scripts
Appearently, there were some bad codes appearing for PWA manifest
and PWA worker discovered by Jemma. This is caused by the improper
Minify input data structure (double "Data" depths) when operating
as a non-server mode. It's a critical bug so we have to fix it.

This patch fixes hugo compilation bug for PWA manifest and worker
scripts in hestiaHUGO/ directory.

Co-authored-by: Veil, Jemma <jemma52@proton.me>
Co-authored-by: Galyna, Cory <cory.galyna@zoralab.com>
Co-authored-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>
Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>