hestiaHUGO: disabled auto URL offline caching due to "memory leak"
It appears that Hugo has a problem with its memory management when
using its resources pipe: it caches without any directive. Each
string data sanitization now contributes to a piece of memory cache
when it's not needed. This is bad as it is shown consuming large
amount of occupied but not usable memory similar to "memory leak".
Hence, we need to reduce it and fix it temporarily by not facilitating
the automatic URL offline caching capability.

This patch disables auto URL offline caching due to "memory leak"
in hestiaHUGO/ directory.

Co-authored-by: Shuralyov, Jean <jean.shuralyov@proton.me>
Co-authored-by: Galyna, Cory <cory.galyna@zoralab.com>
Co-authored-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>
Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>