hestiaHUGO: decontaminated debugger from page assets
We want to make sure that the CSS and JS processed assets output
are indeed page data. Right now, the debugger's assets are
contaminating them as 1st component which is not desirable since
users will copy-paste off the debugger console. Hence, we need to
decontaminate it and debugger has its own delivery pipeline.
Let's do this.

This patch decontaminate debugger from page assets in hestiaHUGO/
directory.

Co-authored-by: Shuralyov, Jean <jean.shuralyov@proton.me>
Co-authored-by: Galyna, Cory <cory.galyna@zoralab.com>
Co-authored-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>
Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>