hestiaHUGO: permit Components and Assets for custom location sourcing
The current configurations permit the content of the page be located
in custom location but not its associated styling assets and components.
This is a required feature in order to render things properly. Hence,
let's do this.

This patch permits Components and Assets to source from custom locations
in hestiaHUGO/ directory.

Co-authored-by: Shuralyov, Jean <jean.shuralyov@proton.me>
Co-authored-by: Galyna, Cory <cory.galyna@zoralab.com>
Co-authored-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>
Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>