sites: added responsive HTML for zoralabIFRAME
Usually, iframe has fixed width and height and it's quite impossible
to set to for responsive behavior based on windows size. However,
after tested and discovered with YouTube iframe embed, we discovered
a scalable and managable way to make it responsively. Hence, let's
document it.

This patch adds responsive HTML code snippet for zoralabIFRAME
GUI component in sites/ directory.

Co-authored-by: Shuralyov, Jean <jean.shuralyov@proton.me>
Co-authored-by: Galyna, Cory <cory.galyna@zoralab.com>
Co-authored-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>
Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>