hestiaHUGO: fixed toggle button vertical positioning bug
There was a UI bug found in zoralabFORM's toggle button (checkbox)
where the vertical alignment simply does not work properly. This is
because that aspect was governed by the CSS Grid and not the elements'
relative positioning. Hence, we need to fix this bug as soon as possible.

This patch fixes toggle button vertical positioning bug in
hestiaHUGO/ directory.

Co-authored-by: Shuralyov, Jean <jean.shuralyov@proton.me>
Co-authored-by: Galyna, Cory <cory.galyna@zoralab.com>
Co-authored-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>
Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>