sites: set /getting-started posts to use linear rendering algorithm
The current /getting-started posts are using recursive methods to
process a simple linear flow blog post. While this is good for
data structuring, it's bad when the page data is huge as it is
currently working at O(n^6) rate. Hence, we need to refactor it
to use linear rendering algorithm instead. Let's do this.

This patch set /getting-started post to use linear rendering algorithm
over its existing recursive algorithm for simple post.

Co-authored-by: Shuralyov, Jean <jean.shuralyov@proton.me>
Co-authored-by: Galyna, Cory <cory.galyna@zoralab.com>
Co-authored-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>
Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>