sites: revamped landing page to be product and customer oriented
The existing landing page was created from the technological perpectives
and its not that straight forward in terms of marketing side of
things. Folks will not understand what is this project about. Hence,
we need to revamp it from the customer perpectives and walk through
the reading journey. Let's do this.

This patch revamps landing page to be product and customer oriented
in sites/ directory.

Co-authored-by: Sibertius <sibertius@gmail.com>
Co-authored-by: Shuralyov, Jean <jean.shuralyov@proton.me>
Co-authored-by: Galyna, Cory <cory.galyna@zoralab.com>
Co-authored-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>
Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>